update video calibration
