+++ src/abb_irb120/Camera_prepare/Camera_calibration/Command Used for Test
@@ -1,0 +1,3 @@
+Command used for test:
+g++ -o video_undistorted video_undistorted.cpp `pkg-config opencv --cflags --libs`
+g++ -o camera_calibration camera_calibration.cpp `pkg-config opencv --cflags --libs`
