+++ src/abb_irb120/Camera_prepare/Camera_calibration/VID5.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0"?>
+<opencv_storage>
+<images>
+images/CameraCalibraation/VID5/xx1.jpg
+images/CameraCalibraation/VID5/xx2.jpg
+images/CameraCalibraation/VID5/xx3.jpg
+images/CameraCalibraation/VID5/xx4.jpg
+images/CameraCalibraation/VID5/xx5.jpg
+images/CameraCalibraation/VID5/xx6.jpg
+images/CameraCalibraation/VID5/xx7.jpg
+images/CameraCalibraation/VID5/xx8.jpg
+</images>
+</opencv_storage>
