+++ src/abb_irb120/Camera_prepare/Camera_test/noimage.launch
@@ -1,0 +1,6 @@
+<launch>
+  <node name="image_view" pkg="image_view" type="image_view" respawn="false" output="screen">
+    <remap from="image" to="/CalibratedVideo/image_fine"/>
+    <param name="autosize" value="false" />
+  </node>
+</launch>
