+++ src/abb_irb120/Camera_prepare/Camera_test/noimage.launch~
@@ -1,0 +1,17 @@
+<launch>
+  <node name="usb_cam" pkg="usb_cam" type="usb_cam_node" output="screen" >
+    <param name="video_device" value="/dev/video0" />
+    <param name="image_width" value="1280" />
+    <param name="image_height" value="720" />
+    <param name="pixel_format" value="yuyv" />
+    <param name="camera_frame_id" value="usb_cam" />
+    <param name="io_method" value="mmap"/>
+    <param name="autofocus" value="true"/>
+    <param name="framerate" value="30"/>
+    <param name="brightness" value="20"/>
+  </node>
+  <node name="image_view" pkg="image_view" type="image_view" respawn="false" output="screen">
+    <remap from="image" to="/CalibratedVideo/image_fine"/>
+    <param name="autosize" value="false" />
+  </node>
+</launch>
