+++ src/abb_irb120/Camera_prepare/Command Used for Test~
@@ -1,0 +1,6 @@
+Command used for test:
+g++ -o ttt perception_node.cpp `pkg-config opencv --cflags --libs`
+
+Command used for Sublime:
+subl
+
