--- src/abb_irb120/Camera_prepare_notROS/Camera_test/Command Used for Test
@@ -1,2 +1,0 @@
-Command used for test:
-g++ -o ttt perception_node.cpp `pkg-config opencv --cflags --libs`
