--- src/abb_irb120/Camera_prepare_notROS/Camera_test/lifecam.launch
@@ -1,10 +1,0 @@
-<launch>
-  <node name="usb_cam" pkg="usb_cam" type="usb_cam_node" output="screen" >
-    <param name="video_device" value="/dev/video0" />
-    <param name="image_width" value="1280" />
-    <param name="image_height" value="720" />
-    <param name="pixel_format" value="yuyv" />
-    <param name="camera_frame_id" value="usb_cam" />
-    <param name="io_method" value="mmap"/>
-  </node>
-</launch>
