--- src/abb_irb120/Camera_prepare_notROS/Camera_test/usb_cam-test.launch
@@ -1,6 +1,0 @@
-<launch>
-  <node name="image_view" pkg="image_view" type="image_view" respawn="false" output="screen">
-    <remap from="image" to="/usb_cam/image_raw"/>
-    <param name="autosize" value="true" />
-  </node>
-</launch>
