--- src/abb_irb120/Camera_prepare_notROS/Command Used for Test~
@@ -1,6 +1,0 @@
-Command used for test:
-g++ -o ttt perception_node.cpp `pkg-config opencv --cflags --libs`
-
-Command used for Sublime:
-subl
-
