+++ src/abb_irb120/irb120_cam_cali/CMakeLists.txt~
@@ -15,6 +15,7 @@ find_package(catkin REQUIRED COMPONENTS
   std_msgs
   tf
   roslib
+  image_transport
 )
 
 ## System dependencies are found with CMake's conventions
@@ -109,7 +110,7 @@ find_package(catkin REQUIRED COMPONENTS
 catkin_package(
 #  INCLUDE_DIRS include
 #  LIBRARIES irb120_cam_cali
-  CATKIN_DEPENDS OpenCV roscpp rospy sensor_msgs std_msgs roslib
+  CATKIN_DEPENDS roslib
 #  DEPENDS system_lib
   INCLUDE_DIRS include
 )
